Feat/auth

+++ src/features/otp/services/otp.service.ts
@@ -13,9 +13,13 @@ export class OtpService implements IOtpService {
   async createOtp(data: CreateOtpDTO): Promise<Otp> {
     data.validate();
     const existingOtp = await this.repository.findOtpByPhone(data.phone);
-    // if (existingOtp) {
-    //   throw new BadRequestError(`You have already requested an otp for this phone number`);
-    // }
+
+
+    if (existingOtp) {
+      const remainingTime = Math.ceil((existingOtp.expiresAt.getTime() - Date.now()) / (1000 * 60));
+      throw new BadRequestError(`You can request an otp after ${remainingTime} minutes`);
+    }
+
     // const code = Math.floor(100000 + Math.random() * 900000).toString();
     const code = '123456';
 
@@ -30,16 +34,16 @@ export class OtpService implements IOtpService {
     const otp = await this.repository.findOtpByPhone(phone);
 
     if (!otp) {
-      // throw new BadRequestError('Otp not found');
-      return false;
+      throw new BadRequestError('Otp not found');
     }
-    // if (otp.code !== code) {
-    //   throw new BadRequestError('Invalid OTP');
-    // }
-    // if (otp.expiresAt < new Date()) {
-    //   throw new BadRequestError('Otp expired');
-    // }
-    console.log(otp.code, code);
+
+    if (otp.code !== code) {
+      throw new BadRequestError('Invalid OTP');
+    }
+
+    if (otp.expiresAt < new Date()) {
+      throw new BadRequestError('Otp expired');
+    }
 
     await this.repository.deleteOtp(otp.id);
